Afrikaans_Chatbot/requirements.txt
speechrecognition
pyaudio
gtts
google-generativeai
configparser
